move graph helper functions into own file
