+++ web/ui/react-app/src/graph/Graph.tsx
@@ -2,10 +2,10 @@ import $ from 'jquery';
 import React, { PureComponent } from 'react';
 import ReactResizeDetector from 'react-resize-detector';
 
-import { escapeHTML } from '../utils/html';
 import SeriesName from '../SeriesName';
 import { Metric, QueryParams } from '../types/types';
 import { isPresent } from '../utils/func';
+import { normalizeData, getOptions, toHoverColor } from './GraphHelpers';
 
 require('flot');
 require('flot/source/jquery.flot.crosshair');
@@ -14,7 +14,7 @@ require('flot/source/jquery.flot.time');
 require('flot/source/jquery.canvaswrapper');
 require('jquery.flot.tooltip');
 
-interface GraphProps {
+export interface GraphProps {
   data: {
     resultType: string;
     result: Array<{ metric: Metric; values: [number, string][] }>;
@@ -23,7 +23,7 @@ interface GraphProps {
   queryParams: QueryParams | null;
 }
 
-interface GraphSeries {
+export interface GraphSeries {
   labels: { [key: string]: string };
   color: string;
   data: (number | null)[][]; // [x,y][]
@@ -42,190 +42,13 @@ class Graph extends PureComponent<GraphProps, GraphState> {
 
   state = {
     selectedSeriesIndex: null,
-    chartData: this.getData(),
+    chartData: normalizeData(this.props),
   };
-
-  formatValue = (y: number | null): string => {
-    if (y === null) {
-      return 'null';
-    }
-    const absY = Math.abs(y);
-
-    if (absY >= 1e24) {
-      return (y / 1e24).toFixed(2) + 'Y';
-    } else if (absY >= 1e21) {
-      return (y / 1e21).toFixed(2) + 'Z';
-    } else if (absY >= 1e18) {
-      return (y / 1e18).toFixed(2) + 'E';
-    } else if (absY >= 1e15) {
-      return (y / 1e15).toFixed(2) + 'P';
-    } else if (absY >= 1e12) {
-      return (y / 1e12).toFixed(2) + 'T';
-    } else if (absY >= 1e9) {
-      return (y / 1e9).toFixed(2) + 'G';
-    } else if (absY >= 1e6) {
-      return (y / 1e6).toFixed(2) + 'M';
-    } else if (absY >= 1e3) {
-      return (y / 1e3).toFixed(2) + 'k';
-    } else if (absY >= 1) {
-      return y.toFixed(2);
-    } else if (absY === 0) {
-      return y.toFixed(2);
-    } else if (absY < 1e-23) {
-      return (y / 1e-24).toFixed(2) + 'y';
-    } else if (absY < 1e-20) {
-      return (y / 1e-21).toFixed(2) + 'z';
-    } else if (absY < 1e-17) {
-      return (y / 1e-18).toFixed(2) + 'a';
-    } else if (absY < 1e-14) {
-      return (y / 1e-15).toFixed(2) + 'f';
-    } else if (absY < 1e-11) {
-      return (y / 1e-12).toFixed(2) + 'p';
-    } else if (absY < 1e-8) {
-      return (y / 1e-9).toFixed(2) + 'n';
-    } else if (absY < 1e-5) {
-      return (y / 1e-6).toFixed(2) + 'µ';
-    } else if (absY < 1e-2) {
-      return (y / 1e-3).toFixed(2) + 'm';
-    } else if (absY <= 1) {
-      return y.toFixed(2);
-    }
-    throw Error("couldn't format a value, this is a bug");
-  };
-
-  getOptions(): jquery.flot.plotOptions {
-    return {
-      grid: {
-        hoverable: true,
-        clickable: true,
-        autoHighlight: true,
-        mouseActiveRadius: 100,
-      },
-      legend: {
-        show: false,
-      },
-      xaxis: {
-        mode: 'time',
-        showTicks: true,
-        showMinorTicks: true,
-        timeBase: 'milliseconds',
-      },
-      yaxis: {
-        tickFormatter: this.formatValue,
-      },
-      crosshair: {
-        mode: 'xy',
-        color: '#bbb',
-      },
-      tooltip: {
-        show: true,
-        cssClass: 'graph-tooltip',
-        content: (_, xval, yval, { series }): string => {
-          const { labels, color } = series;
-          return `
-            <div class="date">${new Date(xval).toUTCString()}</div>
-            <div>
-              <span class="detail-swatch" style="background-color: ${color}" />
-              <span>${labels.__name__ || 'value'}: <strong>${yval}</strong></span>
-            <div>
-            <div class="labels mt-1">
-              ${Object.keys(labels)
-                .map(k =>
-                  k !== '__name__' ? `<div class="mb-1"><strong>${k}</strong>: ${escapeHTML(labels[k])}</div>` : ''
-                )
-                .join('')}
-            </div>
-          `;
-        },
-        defaultTheme: false,
-        lines: true,
-      },
-      series: {
-        stack: this.props.stacked,
-        lines: {
-          lineWidth: this.props.stacked ? 1 : 2,
-          steps: false,
-          fill: this.props.stacked,
-        },
-        shadowSize: 0,
-      },
-    };
-  }
-
-  // This was adapted from Flot's color generation code.
-  getColors() {
-    const colorPool = ['#edc240', '#afd8f8', '#cb4b4b', '#4da74d', '#9440ed'];
-    const colorPoolSize = colorPool.length;
-    let variation = 0;
-    return this.props.data.result.map((_, i) => {
-      // Each time we exhaust the colors in the pool we adjust
-      // a scaling factor used to produce more variations on
-      // those colors. The factor alternates negative/positive
-      // to produce lighter/darker colors.
-
-      // Reset the variation after every few cycles, or else
-      // it will end up producing only white or black colors.
-
-      if (i % colorPoolSize === 0 && i) {
-        if (variation >= 0) {
-          variation = variation < 0.5 ? -variation - 0.2 : 0;
-        } else {
-          variation = -variation;
-        }
-      }
-      return $.color.parse(colorPool[i % colorPoolSize] || '#666').scale('rgb', 1 + variation);
-    });
-  }
-
-  getData(): GraphSeries[] {
-    const colors = this.getColors();
-    const { stacked, queryParams } = this.props;
-    const { startTime, endTime, resolution } = queryParams!;
-    return this.props.data.result.map(({ values, metric }, index) => {
-      // Insert nulls for all missing steps.
-      const data = [];
-      let pos = 0;
-
-      for (let t = startTime; t <= endTime; t += resolution) {
-        // Allow for floating point inaccuracy.
-        const currentValue = values[pos];
-        if (values.length > pos && currentValue[0] < t + resolution / 100) {
-          data.push([currentValue[0] * 1000, this.parseValue(currentValue[1])]);
-          pos++;
-        } else {
-          // TODO: Flot has problems displaying intermittent "null" values when stacked,
-          // resort to 0 now. In Grafana this works for some reason, figure out how they
-          // do it.
-          data.push([t * 1000, stacked ? 0 : null]);
-        }
-      }
-
-      return {
-        labels: metric !== null ? metric : {},
-        color: colors[index].toString(),
-        data,
-        index,
-      };
-    });
-  }
-
-  parseValue(value: string) {
-    const val = parseFloat(value);
-    if (isNaN(val)) {
-      // "+Inf", "-Inf", "+Inf" will be parsed into NaN by parseFloat(). They
-      // can't be graphed, so show them as gaps (null).
-
-      // TODO: Flot has problems displaying intermittent "null" values when stacked,
-      // resort to 0 now. In Grafana this works for some reason, figure out how they
-      // do it.
-      return this.props.stacked ? 0 : null;
-    }
-    return val;
-  }
 
   componentDidUpdate(prevProps: GraphProps) {
-    if (prevProps.data !== this.props.data || prevProps.stacked !== this.props.stacked) {
-      this.setState({ selectedSeriesIndex: null, chartData: this.getData() }, this.plot);
+    const { data, stacked } = this.props;
+    if (prevProps.data !== data || prevProps.stacked !== stacked) {
+      this.setState({ selectedSeriesIndex: null, chartData: normalizeData(this.props) }, this.plot);
     }
   }
 
@@ -239,7 +62,7 @@ class Graph extends PureComponent<GraphProps, GraphState> {
     }
     this.destroyPlot();
 
-    this.$chart = $.plot($(this.chartRef.current), this.state.chartData, this.getOptions());
+    this.$chart = $.plot($(this.chartRef.current), this.state.chartData, getOptions(this.props.stacked));
   };
 
   destroyPlot = () => {
@@ -258,7 +81,9 @@ class Graph extends PureComponent<GraphProps, GraphState> {
   handleSeriesSelect = (index: number) => () => {
     const { selectedSeriesIndex, chartData } = this.state;
     this.plotSetAndDraw(
-      selectedSeriesIndex === index ? chartData.map(this.toHoverColor(index)) : chartData.slice(index, index + 1)
+      selectedSeriesIndex === index
+        ? chartData.map(toHoverColor(index, this.props.stacked))
+        : chartData.slice(index, index + 1)
     );
     this.setState({ selectedSeriesIndex: selectedSeriesIndex === index ? null : index });
   };
@@ -268,7 +93,7 @@ class Graph extends PureComponent<GraphProps, GraphState> {
       cancelAnimationFrame(this.rafID);
     }
     this.rafID = requestAnimationFrame(() => {
-      this.plotSetAndDraw(this.state.chartData.map(this.toHoverColor(index)));
+      this.plotSetAndDraw(this.state.chartData.map(toHoverColor(index, this.props.stacked)));
     });
   };
 
@@ -276,25 +101,6 @@ class Graph extends PureComponent<GraphProps, GraphState> {
     cancelAnimationFrame(this.rafID);
     this.plotSetAndDraw();
   };
-
-  getHoverColor = (color: string, opacity: number) => {
-    const { r, g, b } = $.color.parse(color);
-    if (!this.props.stacked) {
-      return `rgba(${r}, ${g}, ${b}, ${opacity})`;
-    }
-    /*
-      Unfortunetly flot doesn't take into consideration
-      the alpha value when adjusting the color on the stacked series.
-      TODO: find better way to set the opacity.
-    */
-    const base = (1 - opacity) * 255;
-    return `rgb(${Math.round(base + opacity * r)},${Math.round(base + opacity * g)},${Math.round(base + opacity * b)})`;
-  };
-
-  toHoverColor = (index: number) => (series: GraphSeries, i: number) => ({
-    ...series,
-    color: this.getHoverColor(series.color, i !== index ? 0.3 : 1),
-  });
 
   render() {
     const { selectedSeriesIndex, chartData } = this.state;
